change testcase 2517374 and 2517381 to extends AbstractMsalUiTest class and also move them to msalonly.ltw package since they're not broker test cases
